requirements.txt
altair
pandas
psycopg2-binary
sqlalchemy
requests
